Add response caching
Responses from the server are cached using `Dataloader`. Caches survive
only as long as the GraphQL query, so there's no risk of accidental
stale data.

This makes the `scenes {lights {name}}` query case not only possible,
but quick. It could be improved further by requesting every light ahead
of time and priming the cache, but the performance isn't necessary yet.